split ExecutionNode.h
